# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"UserDto": {
				"description": "Detail uživatele",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "ID uživatele",
						"example": 1
					},
					"firstName": {
						"type": "string",
						"description": "Jméno uživatele",
						"example": "Jan"
					},
					"lastName": {
						"type": "string",
						"description": "Příjmení uživatele",
						"example": "Novák"
					},
					"email": {
						"type": "string",
						"description": "Email uživatele",
						"example": "jan.novak@example.com",
						"format": "email"
					},
					"phone": {
						"type": "string",
						"description": "Telefonní číslo uživatele",
						"example": "+420123456789"
					},
					"address": {
						"type": "string",
						"description": "Adresa uživatele",
						"example": "Pražská 123, Praha"
					},
					"isActive": {
						"type": "boolean",
						"description": "Je uživatel aktivní?",
						"example": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "Datum vytvoření záznamu"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "Datum poslední aktualizace záznamu"
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"email",
					"isActive",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateUserDto": {
				"description": "Data pro vytvoření uživatele",
				"properties": {
					"firstName": {
						"type": "string",
						"description": "Jméno uživatele",
						"example": "Jan"
					},
					"lastName": {
						"type": "string",
						"description": "Příjmení uživatele",
						"example": "Novák"
					},
					"email": {
						"type": "string",
						"description": "Email uživatele",
						"example": "jan.novak@example.com",
						"format": "email"
					},
					"password": {
						"type": "string",
						"description": "Heslo uživatele",
						"example": "Password123!",
						"minLength": 8
					},
					"phone": {
						"type": "string",
						"description": "Telefonní číslo uživatele",
						"example": "+420123456789",
						"pattern": "^[+]?[(]?[0-9]{3}[)]?[-\\s.]?[0-9]{3}[-\\s.]?[0-9]{4,6}$"
					},
					"address": {
						"type": "string",
						"description": "Adresa uživatele",
						"example": "Pražská 123, Praha"
					},
					"isActive": {
						"type": "boolean",
						"description": "Je uživatel aktivní?",
						"example": true,
						"default": true
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUserDto": {
				"description": "Data pro aktualizaci uživatele",
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"description": "Entita reprezentující uživatele v systému",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "Unikátní identifikátor uživatele"
					},
					"firstName": {
						"type": "string",
						"description": "Křestní jméno uživatele"
					},
					"lastName": {
						"type": "string",
						"description": "Příjmení uživatele"
					},
					"email": {
						"type": "string",
						"description": "Emailová adresa uživatele (unikátní)"
					},
					"password": {
						"type": "string",
						"description": "Heslo uživatele (nezahrnuto ve výchozích dotazech)"
					},
					"isActive": {
						"type": "boolean",
						"description": "Indikátor, zda je uživatel aktivní"
					},
					"phone": {
						"type": "string",
						"description": "Telefonní číslo uživatele"
					},
					"address": {
						"type": "string",
						"description": "Adresa uživatele"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "Datum vytvoření záznamu"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "Datum poslední aktualizace záznamu"
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"email",
					"password",
					"isActive",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EntityType": {
				"description": "Typ entity - určuje, zda jde o model nebo proceduru",
				"enum": [
					"MODEL",
					"PROCEDURE"
				],
				"type": "string"
			},
			"OperationType": {
				"description": "Výčtový typ pro operace, které lze na entitách provádět",
				"enum": [
					"SELECT",
					"INSERT",
					"UPDATE",
					"DELETE",
					"EXECUTE"
				],
				"type": "string"
			},
			"Role": {
				"description": "Model pro definici rolí v systému\nRole sdružuje sadu oprávnění, která lze přiřadit uživatelům",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "Unikátní identifikátor role"
					},
					"code": {
						"type": "string",
						"description": "Kód role - unikátní identifikátor používaný v kódu\nPříklady: ADMIN, USER, EDITOR, atd."
					},
					"name": {
						"type": "string",
						"description": "Název role - lidsky čitelný popis role"
					},
					"description": {
						"type": "string",
						"description": "Popis role a jejího účelu"
					},
					"isSystem": {
						"type": "boolean",
						"description": "Příznak, zda je role systémová\nSystémové role nelze odstranit ani upravovat standardními operacemi"
					},
					"isActive": {
						"type": "boolean",
						"description": "Příznak, zda je role aktivní"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/Permission"
						},
						"type": "array",
						"description": "Seznam oprávnění přiřazených této roli"
					},
					"users": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array",
						"description": "Seznam uživatelů s touto rolí"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "Datum vytvoření záznamu"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "Datum poslední aktualizace záznamu"
					}
				},
				"required": [
					"id",
					"code",
					"name",
					"isSystem",
					"isActive",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Permission": {
				"description": "Model pro definici oprávnění na entitách\nOprávnění popisuje, jaké operace může uživatel s určitou rolí provádět na dané entitě",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "Unikátní identifikátor oprávnění"
					},
					"entityName": {
						"type": "string",
						"description": "Název cílové entity, ke které se oprávnění vztahuje (např. \"users\", \"products\")"
					},
					"entityType": {
						"$ref": "#/components/schemas/EntityType",
						"description": "Typ entity - MODEL (tabulka) nebo PROCEDURE (procedura)"
					},
					"operation": {
						"$ref": "#/components/schemas/OperationType",
						"description": "Operace, kterou oprávnění umožňuje (SELECT, INSERT, UPDATE, DELETE, EXECUTE)"
					},
					"description": {
						"type": "string",
						"description": "Popis oprávnění"
					},
					"condition": {
						"type": "string",
						"description": "Podmínka pro filtrování záznamů (např. \"id = :userId\")\nMůže být použito pro omezení přístupu pouze k určitým záznamům"
					},
					"roles": {
						"items": {
							"$ref": "#/components/schemas/Role"
						},
						"type": "array",
						"description": "Role, ke kterým je toto oprávnění přiřazeno"
					},
					"isActive": {
						"type": "boolean",
						"description": "Příznak, zda je oprávnění aktivní"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "Datum vytvoření záznamu"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "Datum poslední aktualizace záznamu"
					}
				},
				"required": [
					"id",
					"entityName",
					"entityType",
					"operation",
					"isActive",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserRole": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"roleId": {
						"type": "number",
						"format": "double"
					},
					"roleName": {
						"type": "string"
					},
					"isActive": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					}
				},
				"required": [
					"userId",
					"roleId",
					"roleName",
					"isActive",
					"createdAt",
					"updatedAt",
					"user",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AssignRoleRequest": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"roleId": {
						"type": "number",
						"format": "double"
					},
					"roleName": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"roleId",
					"roleName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateRoleRequest": {
				"properties": {
					"roleName": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"isActive": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "express-oracle-api",
		"version": "1.0.0",
		"description": "REST API s Express.js, TypeORM, Zod a Oracle",
		"contact": {}
	},
	"paths": {
...
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalPages": {
											"type": "number",
											"format": "double"
										},
										"limit": {
											"type": "number",
											"format": "double"
										},
										"page": {
											"type": "number",
											"format": "double"
										},
										"total": {
											"type": "number",
											"format": "double"
										},
										"data": {
											"items": {
												"$ref": "#/components/schemas/UserDto"
											},
											"type": "array"
										}
									},
									"required": [
										"totalPages",
										"limit",
										"page",
										"total",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Získá seznam všech uživatelů s filtrováním a stránkováním",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"description": "Číslo stránky",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Počet záznamů na stránku",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Filtr podle jména",
						"in": "query",
						"name": "firstName",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filtr podle příjmení",
						"in": "query",
						"name": "lastName",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filtr podle emailu",
						"in": "query",
						"name": "email",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"description": "Vytvoří nového uživatele",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Data pro vytvoření uživatele",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDto",
								"description": "Data pro vytvoření uživatele"
							}
						}
					}
				}
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"description": "Získá detail uživatele podle ID",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID uživatele",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"description": "Aktualizuje existujícího uživatele",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID uživatele",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Data pro aktualizaci uživatele",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDto",
								"description": "Data pro aktualizaci uživatele"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Smaže uživatele podle ID",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "ID uživatele",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user-roles/user/{userId}": {
			"get": {
				"operationId": "GetUserRoles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserRole"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Získá všechny role pro konkrétního uživatele",
				"tags": [
					"User Roles"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID uživatele",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user-roles/user/{userId}/role/{roleId}": {
			"get": {
				"operationId": "GetUserRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserRole"
								}
							}
						}
					}
				},
				"description": "Získá detail konkrétní role uživatele\nUkázka práce se složeným primárním klíčem",
				"tags": [
					"User Roles"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID uživatele",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "ID role",
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUserRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserRole"
								}
							}
						}
					}
				},
				"description": "Aktualizuje informace o roli uživatele\nUkázka práce se složeným primárním klíčem",
				"tags": [
					"User Roles"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID uživatele",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "ID role",
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Data pro aktualizaci role",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateRoleRequest",
								"description": "Data pro aktualizaci role"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "RemoveRole",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Odebere roli uživateli\nUkázka použití složeného primárního klíče při mazání",
				"tags": [
					"User Roles"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "ID uživatele",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "ID role",
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user-roles": {
			"post": {
				"operationId": "AssignRole",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserRole"
								}
							}
						}
					}
				},
				"description": "Přiřadí roli uživateli",
				"tags": [
					"User Roles"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Data pro přiřazení role",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AssignRoleRequest",
								"description": "Data pro přiřazení role"
							}
						}
					}
				}
			}
		},
		"/user-roles/user/{userId}/deactivate": {
			"put": {
				"operationId": "DeactivateUserRoles",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Deaktivuje všechny role uživatele\nUkázka hromadné aktualizace",
				"tags": [
					"User Roles"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "ID uživatele",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}